feat: update to discord.js v15

+++ apps/bot/src/commands/slash/purge/match.ts
@@ -6,7 +6,7 @@ import {
 	InteractionContextType,
 	MessageFlags
 } from 'discord-api-types/v10';
-import { Message, PermissionsBitField, type ChatInputCommandInteraction } from 'discord.js';
+import { PermissionsBitField, type ChatInputCommandInteraction } from 'discord.js';
 import { bold, italic } from '@discordjs/formatters';
 
 export default class extends CoreCommand {
@@ -52,32 +52,24 @@ export default class extends CoreCommand {
 		const amount = interaction.options.getInteger('amount') ?? 1;
 		const visible = interaction.options.getBoolean('visible') ?? false;
 
-		const { resource } = await interaction.deferReply({
+		const respond = await interaction.deferReply({
 			...(!visible && { flags: MessageFlags.Ephemeral }),
 			withResponse: true
 		});
-		const reply = resource?.message as Message<true>;
+
+		const reply = respond.resource?.message;
+		if (!reply) return;
 
 		const messages = await interaction.channel?.messages.fetch({ limit: 1e2, cache: true, before: reply.id });
 		const filter = messages?.filter((m) => m.content.includes(content) && !m.pinned);
 
 		const message = Array.from(filter!.values());
+
 		const deleted = await interaction.channel?.bulkDelete(message.slice(0, amount), true);
+		if (!deleted?.length) return interaction.editReply({ content: 'No messages were deleted.' });
 
-		const results = {} as object;
-		for (const [, msg] of deleted!) {
-			const user = msg?.author?.tag;
-			if (!Reflect.get(results, user as string)) Reflect.set(results, user as string, 0);
-			(results as any)[user as string]++;
-		}
-
-		const replies = [
-			`${bold(italic(deleted!.size.toString()))} message(s) have been successfully deleted!`,
-			`${Object.entries(results)
-				.map(([user, size]) => `${italic(`${user}:`)} ${bold(italic(size))}`)
-				.join('\n')}`
-		].join('\n\n');
-
-		return interaction.editReply({ content: replies });
+		return interaction.editReply({
+			content: `${bold(italic(deleted.length.toString()))} message(s) have been successfully deleted!`
+		});
 	}
 }
